Make OidList long, add remark

+++ src/ProSuite.QA.Container/TestContainer/CachedRow.cs
@@ -83,9 +83,9 @@ namespace ProSuite.QA.Container.TestContainer
 			return extent;
 		}
 
-		protected override IList<int> GetOidList()
+		protected override IList<long> GetOidList()
 		{
-			IList<int> oidList = GetOidList(_feature);
+			IList<long> oidList = GetOidList(_feature);
 			return oidList;
 		}
 
@@ -163,7 +163,7 @@ namespace ProSuite.QA.Container.TestContainer
 
 					return cls;
 				}
-			} 
+			}
 
 			public IReadOnlyFeatureClass FeatureClass => (IReadOnlyFeatureClass) ReadOnlyTable;
